Refactors per-platform config test logic
Using wrapper functions to call discrete checks depending on platform,
which is always either Trusty or Xenial. Most of the branching logic was
already present, just cleaning it up a bit in order to make the
functions more readable.

Added a systemd-style check for xvfb under Xenial, since we weren't
inspecting that config on disk.

Accordingly, updates the dev docs to state that we've handled the
config tests.